+++ typescript/sdk/src/token/EvmWarpModule.ts
@@ -1264,6 +1264,7 @@ export class EvmWarpModule extends HyperlaneModule<
           mailbox: actualConfig.mailbox,
           multiProvider: this.multiProvider,
           proxyAdminAddress,
+          rateLimitedSender: this.args.addresses.deployedTokenRoute,
         },
       );
       hookTransactions = result.transactions;