+++ typescript/provider-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperlane-xyz/provider-sdk
 
+## 6.0.1
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@33.1.1
+
 ## 6.0.0
 
 ### Major Changes